Move leaf_function tests into test_leaf_function.py (1/4)
[ghstack-poisoned]
